webinar-05/requirements.txt
appdirs==1.4.4
boto3==1.13.18
botocore==1.16.18
certifi==2020.4.5.1
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
distlib==0.3.0
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
identify==1.4.17
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mysql-connector==2.2.9
nodeenv==1.3.5
pre-commit==2.4.0
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
urllib3==1.25.9
virtualenv==20.0.21
Werkzeug==1.0.1
zipp==3.1.0
